Add `IterMatcher` to wrap `Node`
This adds a `struct` to wrap the `Node` and make the API a little closer
to what people might expect. In particular, it allows adding other
options to the interface as methods on the `struct` rather than
additional parameters on `render()`.

This adds `disable_clippy()` to make it easy to produce matchers that
Clippy does not evaluate.