Remove FIXME: not going to do it
For some applications, e.g. decoding HTML entities, we could let the
iterator be moved when we return `None`, since we can detect the
situation in the caller.

I don’t think that’s a good idea.

To do this reliably, we would have to deal with the last character being
`'&'`, and generally it adds extra confusing logic for what I suspect to
be a pretty minor gain (if that).