Feature: matcher functions that take slices
This renames the existing `render()` functions to `render_iter()` and
adds a new `render_slice()` function. The slice-based matchers seem to
perform somewhat better than the iterator based ones.

I’m not sure having two versions of the `render()` function is the right
way to go about this from a usability perspective.

Also, `IterMatcher` is clearly the wrong name.