Fix build failures in Docker
For some reason non-fresh builds in Docker lose files when the
`target/` directory is shared with the host machine. This generally
results in an error about missing files at the linking step.

This sets `CARGO_TARGET_DIR=/root/target` so that the target directory
is not shared with the host.

Works around issue [#14].

[#14]: https://github.com/danielparks/matchgen/issues/14