Don’t use _all_ HTML entities in tests
Building matchers for all HTML entities can be a bit much for limited
hosts, e.g. micro VMs or limited Docker containers. This removes most of
the HTML entities; it only keeps entites that are all lowercase and
whose names start with one of `[aglqt]`.

This also fixes #14 (tests fail to build in Docker), so it removes the
existing work-around.

This resets all benchmarks.